consumer/requirements.txt
kafka-python-ng
redis==5.0.1
